added check to my.cnf testing

+++ percona/chef/cookbooks/percona/recipes/configure_cluster_server.rb
@@ -30,7 +30,7 @@ ruby_block "removemyconf" do
   block do
 	require 'date'
     #if File.ctime(percona["main_config_file"]).to_date < Date.today
-    if Date.parse(File.ctime("/etc/mysql/my.cnf").to_s) < Date.today
+    if File.exist?("/etc/mysql/my.cnf") && Date.parse(File.ctime("/etc/mysql/my.cnf").to_s) < Date.today
 		Chef::Log.info("****COE-LOG: Deleting pre-existing my.cnf config file")
 		File.delete(percona["main_config_file"])
 	end
